// src/server/model/pCompany.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PCompany { 
    name?: string;
    companyId?: string;
    description?: string;
    financialYearStart?: Date;
    contactperson?: string;
    address1?: string;
    address2?: string;
    address3?: string;
    pin?: string;
    phone?: string;
    fax?: string;
    mobile?: string;
    email?: string;
    weblink?: string;
    pan?: string;
    state?: string;
    country?: string;
    gstin?: string;
    gstType?: string;
}