// src/server/model/pCompanyInput.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PCompany } from './pCompany';

export interface PCompanyInput { 
    company?: PCompany;
    captcha?: string;
    captchaKey?: string;
    username?: string;
    password?: string;
    product?: string;
    adminName?: string;
    adminMobile?: string;
    adminEmail?: string;
    timeZone?: string;
    isRemote?: boolean;
}